LayoutTests/imported/w3c/web-platform-tests/css/selectors/focus-visible-007-expected.txt

This test checks that using the keyboard in a way that does not move focus still causes :focus-visible matching to trigger.
If the user-agent does not claim to support the :focus-visible pseudo-class then SKIP this test.
Use the mouse to focus the element below that says "Click me first."
If the element has a red outline, then the test result is FAILURE.
Press the SHIFT key.
If the element now has a red background, then the test result is FAILURE.
Use the mouse to click the element below that says "Click me second."
If the element has a green background, the test result is SUCCESS.
Click me first.
Click me second.

FAIL Using keyboard while element is focused should trigger :focus-visible; using mouse without moving focus should not cancel it; moving focus using mouse should cancel it. assert_equals: expected "rgb(59, 153, 252)" but got "rgb(0, 0, 0)"
